Make sure to not put signs outside the document range.
Not sure what causes this, but it is easy to reproduce by switching to
different commits in the CLI. There is some mismatch from the truth from
the cli not matching the buffer in the editor.

There is probably a better fix, but I don't the neovim mechanics well enough.