fix: change_delete sign logic
When there's a changed line with deleted line(s) before or after it,
it was only showing the deleted sign (because the previous line was
available).

This also simplifies the logic to always show the change_delete sign
regardless whether the deleted line is before or after the changed line.
Otherwise, it would be kind of confusing to see `▏▔` for a changed line
with deleted lines after it (would need to show `▏▁` instead which would
necessitate more complex logic).